Updated extract

+++ scripts/fix.py
@@ -109,16 +109,16 @@ def move_cleaned(imaging_dir, subject, run):
 
     source_dir = f"{imaging_dir}/{subject}/run{run}.feat/"
     dest_dir = f"{imaging_dir}/../imaging_data_fix/"
-    if not Path(dest_dir + dest.strip('.gz')).exists():
+    if not Path(dest_dir + dest.strip('.gz')).exists() and Path(source_dir + src).exists():
         try:
             shutil.copyfile(source_dir + src,
                             dest_dir + dest)
             gunzip_cmd = f"gunzip {dest_dir + dest}"
             subprocess.run(gunzip_cmd, shell=True)
         except:
-            print('skipping ' + subject + run)
+            print(f'skipping {subject} {run}')
     else:
-        print(subject + run + 'already exists')
+        print(f'{subject} {run} already exists')
